Remove warnings around vector<T>::size

+++ Feeder.cc
@@ -460,30 +460,30 @@ void Feeder::setPos(int ln,int ch)
 
 void Feeder::printPart(int from, int from_char, int to, int to_char)
 {
-	if(from < 0 || from >= m_lines.size())
+	if(from < 0 || from >= (int)m_lines.size())
 		return;
 
 	auto *ln = &m_lines.at(from);
 	for(int i=0;i<from_char;i++)
 		cerr << ' ';
-	for(int i=from_char;i < ln->size();i++){
+	for(int i=from_char;i < (int)ln->size();i++){
 		cerr << ln->at(i);
 	}
 	cerr << endl;
 
-	if(to > m_lines.size())
+	if(to > (int)m_lines.size())
 		return;
 
 	for(int i=from+1;i<to;i++){
 		cerr << m_lines.at(i) << endl;
 	}
 
-	if(to == m_lines.size())
+	if(to == (int)m_lines.size())
 		return;
 
 	ln = &m_lines.at(to);
 	for(int i=0;i<=to_char;i++){
-		if(i < ln->size())
+		if(i < (int)ln->size())
 			cerr << ln->at(i);
 	}
 	cerr << endl;
